GEP-200 Integrate the new toolbar in geppetto-ui
