+++ geppetto.js/geppetto-ui/src/3d-canvas/toolbar/groups/NavigationGroup3D.tsx
