+++ geppetto.js/geppetto-ui/src/3d-canvas/toolbar/groups/PanLeft3D.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 import { Toolbar3DButton } from "../Toolbar3D";
-import { Canvas3DRootState } from "@metacell/geppetto-meta-ui/3d-canvas/Canvas3D";
+import { Canvas3DRootState } from "../../Canvas3D";
 import * as THREE from "three";
 
 interface PanLeft3DProps {
@@ -8,7 +8,10 @@ interface PanLeft3DProps {
   useTransition?: boolean;
 }
 
-const PanLeft3D: React.FC<PanLeft3DProps> = ({ distance = 0.5, useTransition = true }) => {
+const PanLeft3D: React.FC<PanLeft3DProps> = ({
+  distance = 0.5,
+  useTransition = true,
+}) => {
   const handlePanLeft = (fiber: Canvas3DRootState) => {
     if (fiber?.controls) {
       fiber.controls.truck(-distance, 0, useTransition);
